Thread through TilesetContentOptions
