+++ CHANGES.md
@@ -10,7 +10,7 @@
 ##### Additions :tada:
 
 - Added a constructor overload for `Cesium3DTilesSelection::ITwinRealityDataContentLoaderFactory` to override the iTwin Cesium Reality Data base URL. This makes it possible to connect to alternate servers (e.g., staging, QA, mock servers).
-- Added `CesiumGltfReader::MeshPrimitiveModeOptions` to let `CesiumGltfReader::GltfReader` retopologize glTF primitives with unsupported primitive modes.
+- Added `CesiumGltfReader::MeshPrimitiveModeOptions` to make `CesiumGltfReader::GltfReader` retopologize glTF primitives with unsupported primitive modes.
 
 ### v0.61.0 - 2026-06-01
