+++ Cesium3DTilesSelection/src/ImplicitOctreeLoader.cpp
@@ -146,8 +146,7 @@ CesiumAsync::Future<TileLoadResult> requestTileContent(
     const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor,
     const std::string& tileUrl,
     const std::vector<CesiumAsync::IAssetAccessor::THeader>& requestHeaders,
-    CesiumImage::Ktx2TranscodeTargets ktx2TranscodeTargets,
-    bool applyTextureTransform,
+    const Cesium3DTilesSelection::TilesetContentOptions& contentOptions,
     const CesiumUtility::IntrusivePointer<TilesetSharedAssetSystem>&
         pSharedAssetSystem,
     const glm::dmat4& tileTransform,
@@ -155,8 +154,7 @@ CesiumAsync::Future<TileLoadResult> requestTileContent(
   return pAssetAccessor->get(asyncSystem, tileUrl, requestHeaders)
       .thenInWorkerThread(
           [pLogger,
-           ktx2TranscodeTargets,
-           applyTextureTransform,
+           contentOptions,
            asyncSystem,
            pAssetAccessor = pAssetAccessor,
            pSharedAssetSystem,
@@ -201,9 +199,8 @@ CesiumAsync::Future<TileLoadResult> requestTileContent(
 
             if (converter) {
               // Convert to gltf
-              CesiumGltfReader::GltfReaderOptions gltfOptions;
-              gltfOptions.ktx2TranscodeTargets = ktx2TranscodeTargets;
-              gltfOptions.applyTextureTransform = applyTextureTransform;
+              CesiumGltfReader::GltfReaderOptions gltfOptions =
+                  contentOptions.toGltfReaderOptions();
               if (pSharedAssetSystem) {
                 gltfOptions.pSharedAssetSystem = pSharedAssetSystem;
               }
@@ -342,8 +339,7 @@ ImplicitOctreeLoader::loadTileContent(const TileLoadInput& loadInput) {
       pAssetAccessor,
       tileUrl,
       requestHeaders,
-      contentOptions.ktx2TranscodeTargets,
-      contentOptions.applyTextureTransform,
+      contentOptions,
       loadInput.pSharedAssetSystem,
       tile.getTransform(),
       ellipsoid);
