+++ Cesium3DTilesSelection/src/TilesetContentOptions.cpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <Cesium3DTilesSelection/TilesetContentOptions.h>
+
+namespace Cesium3DTilesSelection {
+
+CesiumGltfReader::GltfReaderOptions
+TilesetContentOptions::toGltfReaderOptions() const {
+  CesiumGltfReader::GltfReaderOptions options;
+  options.ktx2TranscodeTargets = this->ktx2TranscodeTargets;
+  options.applyTextureTransform = this->applyTextureTransform;
+  options.primitiveModeOptions = this->primitiveModeOptions;
+}
+
+} // namespace Cesium3DTilesSelection
